userBackEnd/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
pyjwt
bcrypt
python-multipart